ENH: std::cerr redirected while concatenating transforms
Reading transforms with 'h5' extensions was printing many error messages. Indeed, transforms are first read as images, and if ITK throws an exception, they are they read as transforms. This allows to gives transforms and images (deformation fields) without having to specify which one is what. However, 'h5' is an extension both for images and transforms. Therefore, when reading 'h5' transforms, the 'h5' image reader was printing on the screen a lot of error messages that were not relevant for the user. Since exceptions are caught, it does not matter if error messages from ITK are printed on the screen. We therefore redirect the error messages to /dev/null (linux) or NUL: (Windows) to hide them.